Заменит работу с данными на работу с моделью в карточке точки
